add standard ::Logger compatible interface for FluentLogger. Extend ::Logger.
