+++ lib/fluent/logger.rb
@@ -19,6 +19,7 @@ module Fluent
   module Logger
     autoload :ConsoleLogger , 'fluent/logger/console_logger'
     autoload :FluentLogger  , 'fluent/logger/fluent_logger'
+    autoload :LevelFluentLogger  , 'fluent/logger/level_fluent_logger'
     autoload :LoggerBase    , 'fluent/logger/logger_base'
     autoload :TestLogger    , 'fluent/logger/test_logger'
     autoload :TextLogger    , 'fluent/logger/text_logger'
